Changed line graph to bar graph.
Also, I think my implemention can support (in theory) an infinite number of data sets. Only issue I see is that there could be issues if two or more sets have identical data, but that may not be an issue unless both data objects refer to the same memory space (I'm not exactly sure how Javascript handles that, but in my code I don't believe that is a problem with the way I have my arrays set up).